Compiled mbedTlsThreading.cpp object must be linked directly to app
Passing this via static library will cause the low-level initialize to
be dropped during linking. To use it as a static library it would have
to be wrapped in a "-Wl,--whole-archive" + "-Wl,--no-whole-archive", but
then CMake deduplicates these flags, as they are also used for
distortos, breaking the whole build.

It is easier to just pass it as an object and attach it as an "interface
source" to mbedtls target.